feat: migrate routes/index.js to TypeScript
